+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v0.22.0
+
+### Added
+
+- Add support for types and fields named equivalent to PHP keywords like `type Switch` or `{ print { subfield } }`.
+
+### Removed
+
+- No longer generate class `TypeConverters`
+
+## v0.21.2
+
+### Fixed
+
+- Always create log file when instantiating `Log` client
+
 ## v0.21.1
 
 ### Changed